Lots of texture loading stuff, temporarily added libpng and zlib stuff for debug output

+++ src/J3D/J3DShapeFactory.cpp
@@ -146,7 +146,7 @@ J3DShape* J3DShapeFactory::Create(bStream::CStream* stream, uint32_t index) {
 	return newShape;
 }
 
-uint32_t J3DShapeFactory::ConvertPosMtxIndexToDrawIndex(bStream::CStream* stream, const J3DShapeInitData& initData, const uint16_t& packetIndex, const uint16_t& value) {
+uint16_t J3DShapeFactory::ConvertPosMtxIndexToDrawIndex(bStream::CStream* stream, const J3DShapeInitData& initData, const uint16_t& packetIndex, const uint16_t& value) {
 	uint32_t index = 0;
 	
 	uint32_t currentStreamPos = stream->tell();
@@ -179,7 +179,7 @@ uint32_t J3DShapeFactory::ConvertPosMtxIndexToDrawIndex(bStream::CStream* stream
 	return index;
 }
 
-uint32_t J3DShapeFactory::GetUseMatrixValue(bStream::CStream* stream, const J3DShapeInitData& initData, const uint16_t& packetIndex) {
+uint16_t J3DShapeFactory::GetUseMatrixValue(bStream::CStream* stream, const J3DShapeInitData& initData, const uint16_t& packetIndex) {
 	uint32_t index = 0;
 
 	uint32_t currentStreamPos = stream->tell();
